requirements.txt
fastapi
uvicorn
numpy
scikit-learn
faiss-cpu
redis
sentence-transformers
pandas